% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/review.R
\name{review}
\alias{review}
\title{Get a Detailed Review of a Data Frame}
\usage{
review(data, show_review = TRUE, show_warnings = TRUE)
}
\arguments{
\item{data}{A data frame or a path to an external file.}

\item{show_review}{Whether to print the results of the review.}

\item{show_warnings}{Whether to print the warnings about anomalies in the
data (e.g. missing values, duplicated rows, outliers, etc.).}
}
\value{
A list with two elements: \code{review} and \code{warnings}.
}
\description{
\code{review} provides a detailed review of a data, similar to the \code{summary}
function, but with more information. \code{review} also allows you to review
data frames from external files (e.g. CSV, Excel, etc.) or URLs. It returns
a list with two elements: \code{review} and \code{warnings}. The \code{review} element
contains a data frame with the following columns:
\itemize{
\item \code{column}: The name of the column.
\item \code{class}: The class of the column.
\item \code{missing_count}: The number of missing values.
\item \code{missing_rate}: The rate of missing values.
\item \code{factor_ordered}: Whether the column is ordered.
\item \code{factor_unique}: The number of unique values.
\item \code{numeric_mean}: The mean of the column.
\item \code{numeric_sd}: The standard deviation of the column.
\item \code{numeric_min}: The minimum value of the column.
\item \code{numeric_q1}: The first quartile of the column.
\item \code{numeric_median}: The median of the column.
\item \code{numeric_q3}: The third quartile of the column.
\item \code{numeric_max}: The maximum value of the column.
\item \code{character_unique}: The number of distinct values.
\item \code{logical_count}: The number of \code{FALSE} and \code{TRUE}
values.
}
The \code{warnings} element is a list of character vectors with the following
elements:
\itemize{
\item \code{duplicates}: The warnings about duplicated rows.
\item \code{missing}: The warnings about missing values.
\item \code{outliers}: The warnings about outliers.
}
}
\examples{
library(nycflights13)
library(palmerpenguins)

df <- data.frame(
  w = c(1, 2, 3, NA, 5, 6, 7, 8, NA, 10),
  x = c(1L, 2L, 3L, 4L, 5L, 6L, 7L, 8L, 9L, 10L),
  y = c("a", "b", NA, "d", "e", "f", "g", "h", "i", "j"),
  z = c(TRUE, FALSE, TRUE, FALSE, TRUE, FALSE, TRUE, FALSE, TRUE, FALSE)
)

review(df) # Review and warnings
review(df, show_review = FALSE) # Only warnings
review(df, show_warnings = FALSE) # Only review
review(df, show_review = FALSE, show_warnings = FALSE) # Nothing

# You can use `review` with the datasets that come with R or with any other
# data frame you have loaded in your environment.
review(iris)
review(mtcars)
review(airquality)

review(airlines)
review(airports)
review(weather)
review(penguins)

# You can also review a data frame from a path or URL.
review("https://raw.githubusercontent.com/ismayc/pnwflights14/master/data/flights.csv")
}
