% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/review_anomalies.R
\name{review_duplicates}
\alias{review_duplicates}
\title{Get the Duplicate Rows in a Data Frame}
\usage{
review_duplicates(df)
}
\arguments{
\item{df}{A data frame.}
}
\value{
A filtered data frame with the duplicate rows.
}
\description{
\code{review_duplicates} takes a data frame and returns a filters it to only
include the duplicate rows.
}
\examples{
df <- data.frame(
  x = c(1, 2, 3, 3, 4, 5, 5, 5),
  y = c(6, 7, 7, 7, 8, 9, 9, 9),
  z = c(11, 11, 13, 13, 14, 15, 15, 15)
)

review_duplicates(df)

# You can also use `review_duplicates` with the data sets that come with R
# or with any other data set that you have loaded in your environment.
review_duplicates(iris)
review_duplicates(mtcars)
review_duplicates(airquality)
}
