% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/review_stats.R
\name{review_factor_counts}
\alias{review_factor_counts}
\title{Get Counts for Factor Columns in a Data Frame}
\usage{
review_factor_counts(df)
}
\arguments{
\item{df}{A data frame.}
}
\value{
A list of data frames containing the counts for each factor column.
These data frames have the following columns:
\itemize{
\item \code{factor}: The factor value.
\item \code{count}: The number of times the factor value appears in the
column.
}
}
\description{
Get Counts for Factor Columns in a Data Frame
}
\examples{
#
library(palmerpenguins)

review_factor_counts(iris)
review_factor_counts(penguins)
}
