% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/review_stats.R
\name{review_factor_stats}
\alias{review_factor_stats}
\title{Get Statistics for Factor Columns in a Data Frame}
\usage{
review_factor_stats(df)
}
\arguments{
\item{df}{A data frame.}
}
\value{
A data frame with the following columns:
\itemize{
\item \code{column}: The name of the column.
\item \code{factor_ordered}: Whether it is ordered or not.
\item \code{factor_unique}: The number of unique values in the column.
}
}
\description{
\code{review_factor_stats} takes a data frame and returns a data frame with
statistics for the factor columns in the data frame. The columns that
don't match the class of the review function are filled with \code{NA}
values.
}
\examples{
df <- data.frame(
  w = c(1, 2, 3),
  x = c(4L, 5L, 6L),
  y = c("a", "b", "c"),
  z = c(TRUE, FALSE, TRUE)
)

review_factor_stats(df)

# You can also use `review_factor_stats` with the data sets that come
# with R or with any other data set that you have loaded in your
# environment.
review_factor_stats(iris)
review_factor_stats(mtcars)
review_factor_stats(airquality)
}
