% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/review_anomalies.R
\name{review_far_outliers}
\alias{review_far_outliers}
\title{Get the Far Outliers Values in Each Numeric Column}
\usage{
review_far_outliers(df)
}
\arguments{
\item{df}{A data frame.}
}
\value{
A list of numeric vectors with the far outliers in each numeric
}
\description{
\code{review_far_outliers} takes a data frame and returns a list of numeric
vectors with the far outliers in each numeric column. The far outliers are
calculated using the interquartile range (IQR) method or commonly known as
Tukey's fences technique with a multiplier of 3.
}
\examples{
library(nycflights13)

review_far_outliers(iris)
review_far_outliers(mtcars)
review_far_outliers(airquality)
review_far_outliers(flights)
}
