% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/review_anomalies.R
\name{review_missing}
\alias{review_missing}
\title{Get the Number of Missing Data in Each Column}
\usage{
review_missing(df)
}
\arguments{
\item{df}{A data frame.}
}
\value{
A data frame with the following columns:
\itemize{
\item \code{column}: The name of the column.
\item \code{missing_count}: The number of missing data in the column.
\item \code{missing_rate}: The rate of missing data in the column.
}
}
\description{
\code{review_missing} takes a data frame and returns a data frame with the number
and rate of missing data per column.
}
\examples{
df <- data.frame(
  x = c(1, 2, 3, NA, 5),
  y = c(6, 7, NA, 9, 10),
  z = c(11, NA, 13, 14, 15)
)

review_missing(df)

# You can also use `review_missing` with the data sets that come with R or
# with any other data set that you have loaded in your environment.
review_missing(iris)
review_missing(mtcars)
review_missing(airquality)
}
