% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/review_stats.R
\name{review_numeric_stats}
\alias{review_numeric_stats}
\title{Get Statistics for Numeric Columns in a Data Frame}
\usage{
review_numeric_stats(df)
}
\arguments{
\item{df}{A data frame.}
}
\value{
A data frame with the following columns:
\itemize{
\item \code{column}: The name of the column.
\item \code{numeric_mean}: The mean of the column.
\item \code{numeric_sd}: The standard deviation of the column.
\item \code{numeric_min}: The minimum value of the column.
\item \code{numeric_q1}: The 1st quartile of the column.
\item \code{numeric_median}: The median of the column.
\item \code{numeric_q3}: The 3rd quartile of the column.
\item \code{numeric_max}: The maximum value of the column.
}
}
\description{
\code{review_numeric_stats} takes a data frame and returns a data frame with
statistics for the numeric columns in the data frame. The columns that
don't match the class of the review function are filled with \code{NA}
values.
}
\examples{
df <- data.frame(
  w = c(1, 2, 3),
  x = c(4L, 5L, 6L),
  y = c("a", "b", "c"),
  z = c(TRUE, FALSE, TRUE)
)

review_numeric_stats(df)

# You can also use `review_numeric_stats` with the data sets that come
# with R or with any other data set that you have loaded in your
# environment.
review_numeric_stats(iris)
review_numeric_stats(mtcars)
review_numeric_stats(airquality)
}
