% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/review_anomalies.R
\name{review_outliers}
\alias{review_outliers}
\title{Get the Outliers Values in Each Numeric Column}
\usage{
review_outliers(df)
}
\arguments{
\item{df}{A data frame.}
}
\value{
A list of numeric vectors with the outliers in each numeric column.
}
\description{
\code{review_outliers} takes a data frame and returns a list of numeric vectors
with the outliers in each numeric column. The outliers are calculated using
the interquartile range (IQR) method or commonly known as Tukey's fences
technique with a multiplier of 1.5.
}
\examples{
library(nycflights13)

review_outliers(iris)
review_outliers(mtcars)
review_outliers(airquality)
review_outliers(flights)
}
