% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/review_stats.R
\name{review_stats}
\alias{review_stats}
\title{Get a Statistical Review of a Data Frame}
\usage{
review_stats(df)
}
\arguments{
\item{df}{A data frame.}
}
\value{
A data frame with the following columns:
\itemize{
\item \code{column}: The name of the column.
\item \code{factor_ordered}: Whether it is ordered or not.
\item \code{factor_unique}: The number of unique values in the column.
\item \code{numeric_mean}: The mean of the column.
\item \code{numeric_sd}: The standard deviation of the column.
\item \code{numeric_min}: The minimum value of the column.
\item \code{numeric_q1}: The 1st quartile of the column.
\item \code{numeric_median}: The median of the column.
\item \code{numeric_q3}: The 3rd quartile of the column.
\item \code{numeric_max}: The maximum value of the column.
\item \code{character_unique}: The number of unique values in the column.
\item \code{logical_count}: The number of \code{FALSE} and \code{TRUE}
values in the column.
}
}
\description{
\code{review_stats} takes a data frame and performs a statistical review of it.
It calls a review function for each class of column and merges the results
into a single data frame. The columns that don't match the class of the
review function are filled with \code{NA} values. At the moment, the
following classes of columns are supported:
\itemize{
\item \code{factor}
\item \code{numeric}
\item \code{character}
\item \code{logical}
}
}
\examples{
df <- data.frame(
  w = c(1, 2, 3),
  x = c(4L, 5L, 6L),
  y = c("a", "b", "c"),
  z = c(TRUE, FALSE, TRUE)
)

review_stats(df)

# You can also use `review_stats` with the data sets that come with R
# or with any other data set that you have loaded in your environment.
review_stats(iris)
review_stats(mtcars)
review_stats(airquality)
}
