% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/review_warnings.R
\name{review_warnings}
\alias{review_warnings}
\title{Get Explicit Warnings About Anomalies in a Data Frame}
\usage{
review_warnings(df)
}
\arguments{
\item{df}{A data frame.}
}
\value{
A list of character vectors containing the warnings about anomalies
in the data.
}
\description{
\code{review_warnings()} takes a data frame and returns a list of character vectors
containing the warnings about anomalies in the data. The names of the list
elements are "duplicates", "missing", and "outliers", each of which contains
a character vector with the warnings about the corresponding anomaly (except
for "duplicates", which is a single string). If there are no warnings about
a particular anomaly, the corresponding list element is \code{NA}. The warnings
are explicit, so they can be printed directly. The warnings are also
informative, so they can be used to improve the data. For example, the
warnings about missing values can be used to decide whether to impute the
missing values or to remove the corresponding rows or columns.
}
\examples{
review_warnings(iris)
review_warnings(mtcars)
review_warnings(airquality)
}
